MAINT: avoid _lib dependency on optimize
